requirements.txt
requests
pyyaml
bs4
cssutils